+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
-- [Issue #379](https://github.com/manheim/terraform-pipeline/issues/379) Support Terraform 0.15+. Added `-chdir` argument.
+- [Issue #379](https://github.com/manheim/terraform-pipeline/issues/379) Support Terraform 0.15. Added `-chdir` argument.
+- [Issue #395](https://github.com/manheim/terraform-pipeline/issues/395) Support Terraform 1.0. Added `-chdir` argument.
 
 # v5.18