+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 - [Issue #422](https://github.com/manheim/terraform-pipeline/issues/422) TerraformPlugin and TerraformPluginVersion should implement TerraformInitCommandPlugin
+
 # v5.18
 
 - [Issue #214](https://github.com/manheim/terraform-pipeline/issues/214) CredentialsPlugin: support more than username/password (Deprecate `withBuildCredentials`)